checking in some outstanding changes
